requirements.txt
kafka-python
oauthlib==1.0.3
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
tweepy
configparser==3.3.0r2
matplotlib==1.5.0
